Relocate validator types
